test(complex-test): implement complex test for usecase2diagram
