+++ .vscode/launch.json
@@ -14,7 +14,7 @@
         "999999",
         "--colors",
         "--recursive",
-        "${workspaceFolder}/test"
+        "${workspaceFolder}/src"
       ],
       "internalConsoleOptions": "openOnSessionStart",
       "skipFiles": [
